Rename substitutePreflight to substituteTailwindPreflight
